pyproject.toml
[project]
name = "llm-bench-cc"
version = "0.1.0"
description = "Optimization pipeline for google/gemma-4-E4B-it (image-to-text)"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "torch>=2.4",
    "transformers>=4.50",
    "accelerate>=0.34",
    "Pillow>=10",
    "datasets>=3.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.18",
    "codecarbon>=2.7",
    "rapidfuzz>=3.9",
    "jiwer>=3.0",
    "nltk>=3.9",
    "tqdm>=4.66",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6"]
quant = [
    "bitsandbytes>=0.43",
    "auto-gptq>=0.7",
    "autoawq>=0.2",
]

[project.scripts]
llm-bench-cc = "llm_bench_cc.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_bench_cc"]

[tool.ruff]
line-length = 100
target-version = "py310"
